Get off `active_record` notification namespace
Double Entry does a great job of instrumenting various metrics by using
`ActiveSupport::Notifications` however it is currently using the
`active_record` namespace for it's events and it does not implement the
same number of arguments that the framework itself expects which causes
issues with third party gems expecting this.

To resolve this, I've updated the namespace used to instead be
`double_entry` which will allow the third party gems to differentiate
between this gem and the rails framework itself.